+++ lib/simdart.dart
@@ -1,10 +1,10 @@
 export 'src/event.dart';
 export 'src/interval.dart';
 export 'src/observable.dart';
-export 'src/resources_context.dart';
-export 'src/resources.dart';
+export 'src/sim_observer.dart';
+export 'src/event_phase.dart';
+export 'src/resources.dart' hide ResourcesFactory, ResourceStore;
 export 'src/simdart.dart' hide SimDartHelper;
-export 'src/simulation_track.dart';
 export 'src/start_time_handling.dart';
 export 'src/sim_result.dart';
 export 'src/sim_num.dart';