+++ lib/src/simdart.dart
@@ -6,16 +6,14 @@ import 'package:meta/meta.dart';
 import 'package:simdart/src/event.dart';
 import 'package:simdart/src/internal/event_action.dart';
 import 'package:simdart/src/internal/repeat_event_action.dart';
-import 'package:simdart/src/internal/resource.dart';
-import 'package:simdart/src/internal/resources_impl.dart';
 import 'package:simdart/src/internal/simdart_interface.dart';
 import 'package:simdart/src/internal/time_action.dart';
 import 'package:simdart/src/interval.dart';
 import 'package:simdart/src/resources.dart';
 import 'package:simdart/src/sim_counter.dart';
 import 'package:simdart/src/sim_num.dart';
+import 'package:simdart/src/sim_observer.dart';
 import 'package:simdart/src/sim_result.dart';
-import 'package:simdart/src/simulation_track.dart';
 import 'package:simdart/src/start_time_handling.dart';
 
 /// Represents a discrete-event simulation engine.
@@ -28,8 +26,6 @@ class SimDart implements SimDartInterface {
   /// - [startTimeHandling]: Determines how to handle events scheduled with a start
   /// time in the past. The default behavior is [StartTimeHandling.throwErrorIfPast].
   ///
-  /// - [includeTracks]: Determines whether simulation tracks should be included in the simulation result.
-  ///
   /// - [seed]: The optional parameter used to initialize the random number generator
   /// for deterministic behavior in the simulation. If provided, it ensures that the
   /// random sequence is reproducible for the same seed value.
@@ -41,21 +37,22 @@ class SimDart implements SimDartInterface {
   SimDart(
       {this.startTimeHandling = StartTimeHandling.throwErrorIfPast,
       int now = 0,
-      this.includeTracks = false,
+      this.observer,
       this.executionPriority = 0,
       int? seed})
       : random = Random(seed),
         _now = now;
 
-  bool _hasRun = false;
+  RunState _runState = RunState.notStarted;
+  RunState get runState => _runState;
+
+  final SimObserver? observer;
 
   final Map<String, SimNum> _numProperties = {};
   final Map<String, SimCounter> _counterProperties = {};
 
-  /// Holds the resources in the simulator.
-  final Map<String, Resource> _resources = {};
-
-  late final Resources resources = ResourcesImpl(this);
+  final ResourceStore _resourceStore = ResourceStore();
+  late final Resources resources = ResourcesFactory.sim(this);
 
   /// The instance of the random number generator used across the simulation.
   /// It is initialized once and reused to improve performance, avoiding the need to
@@ -85,15 +82,6 @@ class SimDart implements SimDartInterface {
 
   int _executionCount = 0;
 
-  /// Determines whether simulation tracks should be included in the simulation result.
-  ///
-  /// When set to `true`, the simulation will collect and return a list of [SimulationTrack]
-  /// objects as part of its result. If set to `false`, the tracks will not be collected,
-  /// and the list will be `null`.
-  ///
-  /// Default: `false`
-  final bool includeTracks;
-
   int? _startTime;
 
   int _duration = 0;
@@ -106,9 +94,7 @@ class SimDart implements SimDartInterface {
   int get now => _now;
   late int _now;
 
-  List<SimulationTrack>? _tracks;
-
-  Completer<void>? _terminator;
+  final Completer<void> _terminator = Completer();
 
   bool _error = false;
 
@@ -117,33 +103,35 @@ class SimDart implements SimDartInterface {
   /// - [until]: The time at which execution should stop. Execution will include events
   ///   scheduled at this time (inclusive). If null, execution will continue indefinitely.
   Future<SimResult> run({int? until}) async {
-    if (_hasRun) {
-      throw StateError('The simulation has already been run.');
+    if (runState != RunState.notStarted) {
+      throw StateError('Simulation has already started.');
     }
 
-    _hasRun = true;
+    _runState = RunState.running;
+    observer?.onStart();
     if (until != null && _now > until) {
       throw ArgumentError('`now` must be less than or equal to `until`.');
     }
     _until = until;
 
-    if (_terminator != null) {
-      return _buildResult();
-    }
-    if (_actions.isEmpty) {
-      _duration = 0;
-      _startTime = 0;
-      return _buildResult();
-    }
     _duration = 0;
-    _startTime = null;
 
-    _terminator = Completer<void>();
-    _scheduleNextAction();
-    await _terminator?.future;
-    _duration = _now - (_startTime ?? 0);
-    _terminator = null;
+    if (_actions.isEmpty) {
+      _startTime = 0;
+    } else {
+      _startTime = null;
+      _scheduleNextAction();
+      await _terminator.future;
+      _duration = _now - (_startTime ?? 0);
+    }
+    _runState = RunState.finished;
     return _buildResult();
+  }
+
+  void stop() {
+    if (!_terminator.isCompleted) {
+      _terminator.complete();
+    }
   }
 
   @override
@@ -214,7 +202,6 @@ class SimDart implements SimDartInterface {
     return SimResult(
         startTime: _startTime ?? 0,
         duration: _duration,
-        tracks: _tracks,
         numProperties: _numProperties,
         counterProperties: _counterProperties);
   }
@@ -242,26 +229,14 @@ class SimDart implements SimDartInterface {
     _actions.add(action);
   }
 
-  void _addTrack({required String eventName, required Status status}) {
-    Map<String, int> resourceUsage = {};
-    for (Resource resource in _resources.values) {
-      resourceUsage[resource.id] = resource.queue.length;
-    }
-    _tracks ??= [];
-    _tracks!.add(SimulationTrack(
-        status: status,
-        name: eventName,
-        time: now,
-        resourceUsage: resourceUsage));
-  }
-
   Future<void> _consumeNextAction() async {
-    if (_error) {
+    if (_error || _terminator.isCompleted) {
       return;
     }
+
     _nextActionScheduled = false;
     if (_actions.isEmpty) {
-      _terminator?.complete();
+      _terminator.complete();
       return;
     }
 
@@ -272,7 +247,7 @@ class SimDart implements SimDartInterface {
       _now = action.start;
       if (_until != null && now > _until!) {
         _startTime ??= now;
-        _terminator?.complete();
+        _terminator.complete();
         return;
       }
     } else if (action.start < now) {
@@ -284,6 +259,8 @@ class SimDart implements SimDartInterface {
     action.execute();
   }
 }
+
+enum RunState { notStarted, running, finished }
 
 typedef StopCondition = bool Function(SimDart sim);
 
@@ -297,24 +274,8 @@ class SimDartHelper {
     sim._addAction(action);
   }
 
-  static Resource? getResource(
-      {required SimDart sim, required String? resourceId}) {
-    return sim._resources[resourceId];
-  }
-
-  static void addResource(
-      {required SimDart sim,
-      required String resourceId,
-      required Resource Function() create}) {
-    sim._resources.putIfAbsent(resourceId, create);
-  }
-
-  static void addSimulationTrack(
-      {required SimDart sim,
-      required String eventName,
-      required Status status}) {
-    sim._addTrack(eventName: eventName, status: status);
-  }
+  static ResourceStore resourceStore({required SimDart sim}) =>
+      sim._resourceStore;
 
   static void scheduleNextAction({required SimDart sim}) {
     sim._scheduleNextAction();
@@ -322,8 +283,10 @@ class SimDartHelper {
 
   static void error({required SimDart sim, required String msg}) {
     sim._error = true;
-    sim._actions.clear();
-    sim._terminator?.completeError(StateError(msg));
-    sim._terminator = null;
+    while (sim._actions.isNotEmpty) {
+      TimeAction action = sim._actions.removeFirst();
+      action.dispose();
+    }
+    sim._terminator.completeError(StateError(msg));
   }
 }